handle int items

+++ mods/vskeysforlocks/src/BlockBehaviors/BlockBehaviorLockableByKey.cs
@@ -26,7 +26,7 @@ namespace vskeysforlocks.src.BlockBehaviors
             BlockEntity block = byPlayer.Entity.World.BlockAccessor.GetBlockEntity(blockSel.Position);
             if (block != null && block is BlockEntityLockableByKey)
             {
-                if (!String.IsNullOrEmpty(((BlockEntityLockableByKey)block).GetKeySerial()) && Convert.ToInt32(((BlockEntityLockableByKey)block).GetKeySerial()) > 9999)
+                if (!String.IsNullOrEmpty(((BlockEntityLockableByKey)block).GetKeySerial()) && Convert.ToInt64(((BlockEntityLockableByKey)block).GetKeySerial()) > 9999)
                 {
                     if (!HasRightKeyInHands(world, byPlayer, ((BlockEntityLockableByKey)block).GetKeySerial()))
                     {
